+++ src/fr/math/minecraft/shared/inventory/items/armor/IronChessplateCraft.java
@@ -13,7 +13,7 @@ public class IronChessplateCraft extends CraftRecipes {
     }
     @Override
     public void fillRecipe() {
-        CraftData data1 = new CraftData(new int[]
+        CraftData data1 = new CraftData(new byte[]
                 {
                         Material.IRON_INGOT.getId(), Material.AIR.getId(), Material.IRON_INGOT.getId(),
                         Material.IRON_INGOT.getId(), Material.IRON_INGOT.getId(), Material.IRON_INGOT.getId(),
