+++ src/fr/math/minecraft/shared/inventory/items/armor/IronHelmetCraft.java
@@ -13,7 +13,7 @@ public class IronHelmetCraft extends CraftRecipes {
     }
     @Override
     public void fillRecipe() {
-        CraftData data1 = new CraftData(new int[]
+        CraftData data1 = new CraftData(new byte[]
                 {
                         Material.IRON_INGOT.getId(), Material.IRON_INGOT.getId(), Material.IRON_INGOT.getId(),
                         Material.IRON_INGOT.getId(), Material.AIR.getId(), Material.IRON_INGOT.getId(),
@@ -21,7 +21,7 @@ public class IronHelmetCraft extends CraftRecipes {
                 }
         );
 
-        CraftData data2 = new CraftData(new int[]
+        CraftData data2 = new CraftData(new byte[]
                 {
                         Material.AIR.getId(), Material.AIR.getId(), Material.AIR.getId(),
                         Material.IRON_INGOT.getId(), Material.IRON_INGOT.getId(), Material.IRON_INGOT.getId(),
