+++ src/fr/math/minecraft/shared/inventory/items/axe/DiamondAxeCraft.java
@@ -13,7 +13,7 @@ public class DiamondAxeCraft extends CraftRecipes {
     }
     @Override
     public void fillRecipe() {
-        CraftData data1 = new CraftData(new int[]
+        CraftData data1 = new CraftData(new byte[]
                 {
                         Material.DIAMOND.getId(), Material.DIAMOND.getId(), Material.AIR.getId(),
                         Material.DIAMOND.getId(), Material.STICK.getId(), Material.AIR.getId(),
@@ -21,7 +21,7 @@ public class DiamondAxeCraft extends CraftRecipes {
                 }
         );
 
-        CraftData data2 = new CraftData(new int[]
+        CraftData data2 = new CraftData(new byte[]
                 {
                         Material.AIR.getId(), Material.DIAMOND.getId(), Material.DIAMOND.getId(),
                         Material.AIR.getId(), Material.STICK.getId(), Material.DIAMOND.getId(),
