+++ src/fr/math/minecraft/shared/inventory/items/sword/WoodenSwordCraft.java
@@ -13,7 +13,7 @@ public class WoodenSwordCraft extends CraftRecipes {
     }
     @Override
     public void fillRecipe() {
-        CraftData data1 = new CraftData(new int[]
+        CraftData data1 = new CraftData(new byte[]
                 {
                         Material.OAK_PLANKS.getId(), Material.AIR.getId(), Material.AIR.getId(),
                         Material.OAK_PLANKS.getId(), Material.AIR.getId(), Material.AIR.getId(),
@@ -21,7 +21,7 @@ public class WoodenSwordCraft extends CraftRecipes {
                 }
         );
 
-        CraftData data2 = new CraftData(new int[]
+        CraftData data2 = new CraftData(new byte[]
                 {
                         Material.AIR.getId(), Material.OAK_PLANKS.getId(), Material.AIR.getId(),
                         Material.AIR.getId(), Material.OAK_PLANKS.getId(), Material.AIR.getId(),
@@ -29,7 +29,7 @@ public class WoodenSwordCraft extends CraftRecipes {
                 }
         );
 
-        CraftData data3 = new CraftData(new int[]
+        CraftData data3 = new CraftData(new byte[]
                 {
                         Material.AIR.getId(), Material.AIR.getId(), Material.OAK_PLANKS.getId(),
                         Material.AIR.getId(), Material.AIR.getId(), Material.OAK_PLANKS.getId(),
